move kernel infos into separate module
